requirements.txt
APScheduler
channels
channels_redis
Django
django_apscheduler
djangorestframework_simplejwt
django_filter
djangorestframework
django-cors-headers
django_mirage_field
django-redis
django-redis-cache
django-mptt
djangorestframework-recursive
requests
aead
pyvim
vmwc
pysmb
mirage
pyOpenSSL
pyVmomi
PyJWT
psycopg2-binary
exchangelib
openpyxl
jpype1
impacket
requests
pytz